requirements.txt
ephem==3.7.6.0
pyenchant==1.6.8
python-telegram-bot==5.3.0
